開発元コードの名称変更
